fix: DF-380 - Move factory methods into separate file to resolve component circular imports
